added dummy spindle_voltage hal component so older machines can assume it's there
spindle_voltage component is created on machines that use the i2c DAC to
control the spindle voltage. It also monitors the clock pin to meaure the
actual spindle speed. Older machines don't have the i2c chip or the clock
pin hooked up, but we want to be able to seamlessly use it as if it were
(it simply returns 0 if the hardware isn't in place).
SOFT-64